#4 - extracting components to separate files
