+++ components/SessionsPageHeader.tsx
@@ -1,0 +1,12 @@
+import { Center } from "@chakra-ui/layout";
+import Logo from "./Logo";
+
+function SessionsPageHeader() {
+  return (
+    <Center>
+      <Logo size="md" />
+    </Center>
+  );
+}
+
+export default SessionsPageHeader;
